+++ src/status-bar-items/verified-status-bar-item.ts
@@ -1,9 +1,8 @@
 import { IVerifyResult } from "@veramo/core";
 import * as vscode from "vscode";
 import { getVeramo } from "../veramo";
+import { generateCIDForString, getIssuerDID } from '../utils';
 import { posix } from 'path';
-import pkg from 'blakejs';
-const { blake2bHex } = pkg;
 import yaml from 'yaml';
 import matter from 'gray-matter';
 
@@ -11,7 +10,7 @@ const verifiedStatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBar
 export { verifiedStatusBarItem };
 
 export async function updateVerifiedStatusBarItem() {
-  verifiedStatusBarItem.text = 'Verifying...';
+  verifiedStatusBarItem.hide();
   verifiedStatusBarItem.color = new vscode.ThemeColor('foreground');
   verifiedStatusBarItem.command = undefined;
   const editor = vscode.window.activeTextEditor;
@@ -19,7 +18,7 @@ export async function updateVerifiedStatusBarItem() {
   const language = editor?.document.languageId;
 
   if (text && editor) {
-    let cid = blake2bHex(text);
+    let cid = await generateCIDForString(text);
     const veramo = getVeramo();
     let credential;
     let result: IVerifyResult | undefined = undefined;
@@ -27,9 +26,11 @@ export async function updateVerifiedStatusBarItem() {
       switch(language) {
         case 'json':
           credential = JSON.parse(text);
+          verifiedStatusBarItem.text = 'Verifying...';
+          verifiedStatusBarItem.show();
           result = await veramo.verifyCredential({credential});
           if (result?.verified) {
-            verifiedStatusBarItem.text = `$(check) Signed by ${result.issuer} (${veramo.context.name})`;
+            verifiedStatusBarItem.text = `$(check) Signed by ${getIssuerDID(credential)} (${veramo.context.name})`;
             verifiedStatusBarItem.command = 'veramo.verify-credential';
             verifiedStatusBarItem.show();
           } else {
@@ -40,9 +41,11 @@ export async function updateVerifiedStatusBarItem() {
           break;
         case 'yaml':
           credential = yaml.parse(text);
+          verifiedStatusBarItem.text = 'Verifying...';
+          verifiedStatusBarItem.show();
           result = await veramo.verifyCredential({credential});
           if (result?.verified) {
-            verifiedStatusBarItem.text = `$(check) Signed by ${result.issuer} (${veramo.context.name})`;
+            verifiedStatusBarItem.text = `$(check) Signed by ${getIssuerDID(credential)} (${veramo.context.name})`;
             verifiedStatusBarItem.command = 'veramo.verify-credential';
             verifiedStatusBarItem.show();
           } else {
@@ -55,17 +58,19 @@ export async function updateVerifiedStatusBarItem() {
           const parsed = matter(text);
           if (parsed.data && parsed.content) {
             credential = parsed.data as any;
-            cid = blake2bHex(parsed.content);
-            result = await veramo.verifyCredential({credential});
-            if (result.verified && result.verifiableCredential?.credentialSubject?.cid === cid) {
-              verifiedStatusBarItem.text = `$(check) Signed by ${result.issuer} (${veramo.context.name})`;
+            cid = await generateCIDForString(parsed.content);
+            verifiedStatusBarItem.text = 'Verifying...';
+            verifiedStatusBarItem.show();
+              result = await veramo.verifyCredential({credential});
+            if (result?.verified && result?.verifiableCredential?.credentialSubject?.cid === cid) {
+              verifiedStatusBarItem.text = `$(check) Signed by ${getIssuerDID(credential)} (${veramo.context.name})`;
               verifiedStatusBarItem.command = {
                 title: 'Verify credential',
                 command: 'veramo.verify-credential',
                 arguments: [{str: JSON.stringify(parsed.data)}]
               };
               verifiedStatusBarItem.show();
-            } else if (result.verifiableCredential?.credentialSubject?.cid && result.verifiableCredential?.credentialSubject?.cid !== cid) {
+            } else if (result?.verifiableCredential?.credentialSubject?.cid && result?.verifiableCredential?.credentialSubject?.cid !== cid) {
               verifiedStatusBarItem.text = `$(error) CID don't match (${veramo.context.name})`;
               verifiedStatusBarItem.color = new vscode.ThemeColor('errorForeground');
               verifiedStatusBarItem.command = {
@@ -86,7 +91,7 @@ export async function updateVerifiedStatusBarItem() {
           break;
         default:
           if (vscode.workspace.workspaceFolders) {
-            const cid = blake2bHex(text);
+            const cid = await generateCIDForString(text);
             const contextFolder = vscode.workspace.getConfiguration("veramo").get("contextFolder", "context");
             const folderUri = vscode.workspace.workspaceFolders[0].uri;
             const fileUri = folderUri.with({ path: posix.join(folderUri.path, contextFolder, `${cid}.json`) });
@@ -94,15 +99,17 @@ export async function updateVerifiedStatusBarItem() {
             const readData = await vscode.workspace.fs.readFile(fileUri);
             const fileContents = Buffer.from(readData).toString('utf8');
             const path = vscode.workspace.asRelativePath(editor.document.uri);
-  
-            result = await veramo.verifyCredential({credential: JSON.parse(fileContents)});
+            verifiedStatusBarItem.text = 'Verifying...';
+            verifiedStatusBarItem.show();
+            const credential = JSON.parse(fileContents);
+            result = await veramo.verifyCredential({credential});
             if (
               result?.verified 
               && (result.verifiableCredential?.type as string[]).includes('SignedFile')
               && result.verifiableCredential?.credentialSubject.id === cid
               && result.verifiableCredential?.credentialSubject.path === path
             ) {
-              verifiedStatusBarItem.text = `$(check) Signed by ${result.issuer} (${veramo.context.name})`;
+              verifiedStatusBarItem.text = `$(check) Signed by ${getIssuerDID(credential)} (${veramo.context.name})`;
               verifiedStatusBarItem.command = {
                 title: 'Verify credential',
                 command: 'veramo.verify-credential',